+++ src/layout/gsub1.rs
@@ -82,3 +82,59 @@ impl ToLowlevel<GSUBSubtable> for SingleSubst {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::layout::common::{Lookup, LookupFlags};
+    use crate::tables::GSUB::tests::{assert_can_roundtrip, expected_gsub};
+    use crate::tables::GSUB::Substitution;
+    use otspec::btreemap;
+    use std::iter::FromIterator;
+
+    #[test]
+    fn test_gsub1_format1_deser() {
+        let binary_gsub = vec![
+            0x00, 0x01, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x1e, 0x00, 0x2c, 0x00, 0x01, 0x44, 0x46,
+            0x4c, 0x54, 0x00, 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0x00, 0x01,
+            0x00, 0x00, 0x00, 0x01, 0x74, 0x65, 0x73, 0x74, 0x00, 0x08, 0x00, 0x00, 0x00, 0x01,
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x04, 0x00, 0x01, 0x00, 0x00, 0x00, 0x01, 0x00, 0x08,
+            0x00, 0x01, 0x00, 0x06, 0x00, 0x01, 0x00, 0x01, 0x00, 0x02, 0x00, 0x42, 0x00, 0x45,
+        ];
+        let expected = expected_gsub(vec![Lookup {
+            flags: LookupFlags::empty(),
+            mark_filtering_set: None,
+            rule: Substitution::Single(vec![SingleSubst {
+                mapping: btreemap!(
+                    66 => 67,
+                    69 => 70
+                ),
+            }]),
+        }]);
+        assert_can_roundtrip(binary_gsub, &expected);
+    }
+
+    #[test]
+    fn test_gsub1_format2_deser() {
+        let binary_gsub = vec![
+            0x00, 0x01, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x1e, 0x00, 0x2c, 0x00, 0x01, 0x44, 0x46,
+            0x4c, 0x54, 0x00, 0x08, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0x00, 0x01,
+            0x00, 0x00, 0x00, 0x01, 0x74, 0x65, 0x73, 0x74, 0x00, 0x08, 0x00, 0x00, 0x00, 0x01,
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x04, 0x00, 0x01, 0x00, 0x00, 0x00, 0x01, 0x00, 0x08,
+            0x00, 0x02, 0x00, 0x0c, 0x00, 0x03, 0x00, 0x43, 0x00, 0x44, 0x00, 0x49, 0x00, 0x01,
+            0x00, 0x03, 0x00, 0x42, 0x00, 0x43, 0x00, 0x46,
+        ];
+        let expected = expected_gsub(vec![Lookup {
+            flags: LookupFlags::empty(),
+            mark_filtering_set: None,
+            rule: Substitution::Single(vec![SingleSubst {
+                mapping: btreemap!(
+                    66 => 67,
+                    67 => 68,
+                    70 => 73
+                ),
+            }]),
+        }]);
+        assert_can_roundtrip(binary_gsub, &expected);
+    }
+}
