+++ src/tables/GSUB.rs
@@ -188,11 +188,9 @@ pub(crate) fn to_bytes(
 #[cfg(test)]
 pub(crate) mod tests {
     use super::*;
-    use crate::layout::common::{
-        FeatureList, LanguageSystem, LookupFlags, Script, ScriptList, ValueRecord,
-    };
+    use crate::layout::common::{FeatureList, LanguageSystem, LookupFlags, Script, ScriptList};
     use crate::tag;
-    use otspec::{btreemap, valuerecord};
+    use otspec::btreemap;
     use std::collections::BTreeMap;
     use std::iter::FromIterator;
 
@@ -228,50 +226,5 @@ pub(crate) mod tests {
         let mut gsub_data = vec![];
         to_bytes(&gsub, &mut gsub_data, 200).unwrap();
         assert_eq!(gsub_data, binary_gsub);
-    }
-
-    #[test]
-    fn test_gsub1_highlevel_de() {
-        let binary_gsub = vec![
-            0x00, 0x01, 0x00, 0x00, // GSUB 1.0
-            0x00, 0x0a, // scriptlist offset
-            0x00, 0x1e, // featurelist offset
-            0x00, 0x2c, // lookuplist offset
-            /* 0x0a */ 0x00, 0x01, // ScriptList.scriptCount
-            0x44, 0x46, 0x4c, 0x54, // ScriptRecord.scriptTag = DFLT
-            0x00, 0x08, // ScriptRecord.scriptOffset
-            0x00, 0x04, // Script.defaultLangSysOffset
-            0x00, 0x00, // Script.langSysCount
-            0x00, 0x00, // LangSys.lookupOrderOffset
-            0xff, 0xff, // LangSys.requiredFeatureIndex
-            0x00, 0x01, // LangSys.featureIndexCount
-            0x00, 0x00, // LangSys.featureIndices
-            /* 0x1e */ 0x00, 0x01, // FeatureList.featureCount
-            0x74, 0x65, 0x73, 0x74, //FeatureRecord.featureTag = test
-            0x00, 0x08, // FeatureRecord.featureOffset
-            0x00, 0x00, // Feature.featureParamsOffset
-            0x00, 0x01, // Feature.lookupIndexCount
-            0x00, 0x00, // Feature.lookupListIndices
-            /* 0x2c */ 0x00, 0x01, // LookupList.lookupCount
-            0x00, 0x04, // LookupList.lookupOffsets
-            0x00, 0x01, // Lookup.lookupType
-            0x00, 0x00, // Lookup.lookupFlags
-            0x00, 0x01, // Lookup.subtableCount
-            0x00, 0x08, // Lookup.subtableOffsets
-            0x00, 0x01, 0x00, 0x08, 0x00, 0x04, 0x00, 0x23, 0x00, 0x01, 0x00, 0x03, 0x00, 0x25,
-            0x00, 0x30, 0x00, 0x32,
-        ];
-        let expected = expected_gsub(vec![Lookup {
-            flags: LookupFlags::empty(),
-            mark_filtering_set: None,
-            rule: Substitution::Single(vec![SinglePos {
-                mapping: btreemap!(
-                    37 => valuerecord!(xAdvance = 35),
-                    48 => valuerecord!(xAdvance = 35),
-                    50 => valuerecord!(xAdvance = 35)
-                ),
-            }]),
-        }]);
-        assert_can_roundtrip(binary_gsub, &expected);
     }
 }
